renamed DataView to data-view
