+++ sites/geohub/src/components/data-view/AddLayerButton.svelte
