+++ sites/geohub/src/components/data-view/DataCardInfo.svelte
