+++ sites/geohub/src/components/data-view/DataCategoryCard.svelte
