+++ sites/geohub/src/components/data-view/DataStacAssetCard.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import type { AssetOptions, BannerMessage, StacItemFeature } from '$lib/types'
   import Accordion from '$components/controls/Accordion.svelte'
-  import AddLayerButton from '$components/DataView/AddLayerButton.svelte'
+  import AddLayerButton from '$components/data-view/AddLayerButton.svelte'
   import { map, layerList, indicatorProgress, bannerMessages } from '$stores'
   import { MosaicJsonData } from '$lib/MosaicJsonData'
   import { StatusTypes } from '$lib/constants'
