+++ sites/geohub/src/components/data-view/MiniMap.svelte
